cleanup and moved translation bias test to tests_staging
